Make calendar

+++ routes/speech.rb
@@ -6,15 +6,15 @@ class App < Sinatra::Base
   # use comma to separate each status
   # /speeches?status=auditing,confirmed
   get '/speeches' do
-    status = params[:status] || Constants::CONFIRMED + ',' + Constants::FINISHED
+    status = params[:status] || Constants::SPEECH_STATUS::CONFIRMED + ',' + Constants::SPEECH_STATUS::FINISHED
     Speech.where(status: status.gsub(/\s+/, '').split(',')).order(time: :desc).to_json
   end
 
   # retrieve a speech, including applied audiences (not finished) or participants (finished)
   get '/speeches/:speech_id' do
     speech = Speech.find(params[:speech_id])
-    if speech.status == Constants::FINISHED
-      speech.to_json(include: :participants)
+    if speech.status == Constants::SPEECH_STATUS::FINISHED
+      speech.to_json(include: :attendances)
     else
       speech.to_json(include: :audiences)
     end
@@ -24,7 +24,7 @@ class App < Sinatra::Base
   post '/speeches' do
     speech = Speech.new(title: @body['title'], description: @body['description'],
                         user_id: @userid, expected_duration: @body['expected_duration'],
-                        status: Constants::NEW, category: @body['category'])
+                        status: Constants::SPEECH_STATUS::NEW, category: @body['category'])
     speech.save!
     speech.to_json
   end
@@ -34,7 +34,7 @@ class App < Sinatra::Base
   put '/speeches/:speech_id' do
     speech = Speech.find(params[:speech_id])
     self_required! speech.user_id
-    if speech.status != Constants::NEW
+    if speech.status != Constants::SPEECH_STATUS::NEW
       400
     else
       speech.title = @body['title']
@@ -51,8 +51,9 @@ class App < Sinatra::Base
   delete '/speeches/:speech_id' do
     speech = Speech.find(params[:speech_id])
     self_required! speech.user_id
-    if speech.status == Constants::NEW
+    if speech.status == Constants::SPEECH_STATUS::NEW
       speech.destroy!
+      content_type 'text/plain'
       200
     else
       400
@@ -63,10 +64,11 @@ class App < Sinatra::Base
   post '/speeches/:speech_id/upload' do
     speech = Speech.find(params[:speech_id])
     self_required! speech.user_id
-    if speech.status == Constants::CONFIRMED
+    if speech.status == Constants::SPEECH_STATUS::CONFIRMED
       speech.resource_url = @body['resource_url']
+      speech.resource_name = @body['resource_name']
       speech.save!
-      200
+      speech.to_json(include: :audiences)
     else
       400
     end
@@ -76,10 +78,10 @@ class App < Sinatra::Base
   post '/speeches/:speech_id/submit' do
     speech = Speech.find(params[:speech_id])
     self_required! speech.user_id
-    if speech.status == Constants::NEW
-      speech.status = Constants::AUDITING
+    if speech.status == Constants::SPEECH_STATUS::NEW
+      speech.status = Constants::SPEECH_STATUS::AUDITING
       speech.save!
-      200
+      speech.to_json
     else
       400
     end
@@ -88,10 +90,10 @@ class App < Sinatra::Base
   post '/speeches/:speech_id/withdraw' do
     speech = Speech.find(params[:speech_id])
     self_required! speech.user_id
-    if speech.status == Constants::AUDITING || speech.status == Constants::APPROVED
-      speech.status = Constants::NEW
+    if speech.status == Constants::SPEECH_STATUS::AUDITING || speech.status == Constants::SPEECH_STATUS::APPROVED
+      speech.status = Constants::SPEECH_STATUS::NEW
       speech.save!
-      200
+      speech.to_json
     else
       400
     end
@@ -100,11 +102,11 @@ class App < Sinatra::Base
   post '/speeches/:speech_id/approve' do
     admin_required!
     speech = Speech.find(params[:speech_id])
-    if speech.status == Constants::AUDITING
-      speech.status = Constants::APPROVED
+    if speech.status == Constants::SPEECH_STATUS::AUDITING
+      speech.status = Constants::SPEECH_STATUS::APPROVED
       speech.time = @body['time']
       speech.save!
-      200
+      speech.to_json
     else
       400
     end
@@ -113,22 +115,27 @@ class App < Sinatra::Base
   post '/speeches/:speech_id/reject' do
     admin_required!
     speech = Speech.find(params[:speech_id])
-    if speech.status == Constants::AUDITING
-      speech.status = Constants::NEW
+    if speech.status == Constants::SPEECH_STATUS::AUDITING
+      speech.status = Constants::SPEECH_STATUS::NEW
       speech.save!
-      200
+      speech.to_json
     else
       400
     end
   end
   # approved -> confirmed
   post '/speeches/:speech_id/agree' do
+    puts request.cookies
     speech = Speech.find(params[:speech_id])
     self_required! speech.user_id
-    if speech.status == Constants::APPROVED
-      speech.status = Constants::CONFIRMED
-      speech.save!
-      200
+    if speech.status == Constants::SPEECH_STATUS::APPROVED
+      ActiveRecord::Base.transaction do
+        speech.status = Constants::SPEECH_STATUS::CONFIRMED
+        event = CalendarHelper::post_event(request.cookies, @userid, speech.title, speech.description, speech.time, speech.time, @userid)
+        speech.event_id = JSON.parse(event)['id']
+        speech.save!
+      end
+      speech.to_json
     else
       400
     end
@@ -137,10 +144,10 @@ class App < Sinatra::Base
   post '/speeches/:speech_id/disagree' do
     speech = Speech.find(params[:speech_id])
     self_required! speech.user_id
-    if speech.status == Constants::APPROVED
-      speech.status = Constants::AUDITING
+    if speech.status == Constants::SPEECH_STATUS::APPROVED
+      speech.status = Constants::SPEECH_STATUS::AUDITING
       speech.save!
-      200
+      speech.to_json
     else
       400
     end
@@ -149,10 +156,13 @@ class App < Sinatra::Base
   post '/speeches/:speech_id/close' do
     admin_required!
     speech = Speech.find(params[:speech_id])
-    if speech.status == Constants::CONFIRMED
-      speech.status = Constants::CLOSED
-      speech.save!
-      200
+    if speech.status == Constants::SPEECH_STATUS::CONFIRMED
+      ActiveRecord::Base.transaction do
+        CalendarHelper::delete_event(request.cookies, @userid, speech.event_id)
+        speech.status = Constants::SPEECH_STATUS::CLOSED
+        speech.save!
+      end
+      speech.to_json(include: :audiences)
     else
       400
     end
@@ -161,10 +171,33 @@ class App < Sinatra::Base
   post '/speeches/:speech_id/finish' do
     admin_required!
     speech = Speech.find(params[:speech_id])
-    if speech.status == Constants::CONFIRMED
-      speech.status = Constants::FINISHED
-      speech.save!
-      200
+    if speech.status == Constants::SPEECH_STATUS::CONFIRMED
+      participants = @body['participants']
+
+      ActiveRecord::Base.transaction do
+        participants.each{|u|
+          not_add_point = true
+          user = User.find_by_id(u['user_id'])
+          if user.nil?
+            user = User.new(id: u['user_id'], role: Constants::USER_ROLE::USER, point_total: 0, point_available: 0)
+            user.change_point(u['point'] + (u['commented'] ? 1 : 0))
+            user.save!
+            not_add_point = false
+          end
+          unless Attendance.exists?(user_id: u['user_id'], speech_id: params[:speech_id])
+              attendance = Attendance.new(user_id: u['user_id'], speech_id: params[:speech_id],
+                                          role: u['role'], point: u['point'], commented: u['commented'])
+              attendance.save!
+              if not_add_point
+                user.change_point(u['point'] + (u['commented'] ? 1 : 0))
+                user.save!
+              end
+          end
+        }
+        speech.status = Constants::SPEECH_STATUS::FINISHED
+        speech.save!
+        speech.to_json(include: :attendances)
+      end
     else
       400
     end
@@ -176,16 +209,31 @@ class App < Sinatra::Base
 
   # user apply to be an audience
   post '/speeches/:speech_id/audiences' do
-    audience = AudienceRegistration.new(user_id: @userid, speech_id: params[:speech_id])
-    audience.save!
-    audience.to_json
+    speech = Speech.find(params[:speech_id])
+    halt 400 if speech.status != Constants::SPEECH_STATUS::CONFIRMED
+    unless AudienceRegistration.exists?(user_id: @userid, speech_id: params[:speech_id])
+      ActiveRecord::Base.transaction do
+        CalendarHelper::apply(request.cookies, @userid, speech.event_id, @userid)
+        audience = AudienceRegistration.new(user_id: @userid, speech_id: params[:speech_id])
+        audience.save!
+      end
+    end
+    speech.to_json(include: :audiences)
   end
 
   # user withdraw his apply to be an audience
   delete '/speeches/:speech_id/audiences/:user_id' do
     self_required! params[:user_id].to_i
-    AudienceRegistration.destroy_all(user_id: params[:user_id], speech_id: params[:speech_id])
-    200
+    speech = Speech.find(params[:speech_id])
+    halt 400 if speech.status != Constants::SPEECH_STATUS::CONFIRMED
+    registration = AudienceRegistration.where(user_id: params[:user_id], speech_id: params[:speech_id])
+    unless registration.empty?
+      ActiveRecord::Base.transaction do
+        AudienceRegistration.destroy_all(user_id: params[:user_id], speech_id: params[:speech_id])
+        CalendarHelper::withdraw_apply(request.cookies, @userid, speech.event_id, @userid)
+      end
+    end
+    speech.to_json(include: :audiences)
   end
 
   get '/speeches/:speech_id/participants' do
@@ -205,6 +253,7 @@ class App < Sinatra::Base
         user.save!
       end
     end
+    content_type 'text/plain'
     200
   end
 
@@ -221,6 +270,7 @@ class App < Sinatra::Base
         Attendance.destroy_all(user_id: params[:user_id], speech_id: params[:speech_id])
       end
     end
+    content_type 'text/plain'
     200
   end
